refactor(native): split GitOps / RpcDispatcher / GozdApp and reorganize GozdCore into subdirectories
intent(native): 1000+ 行のファイル (GitOps 1979 / RpcDispatcher 1394 / GozdApp 925) を責務単位で見直せる粒度に揃える
decision(gitops): public enum GitOps を namespace のまま extension で 7 ファイルに分割し、status / worktree / branch / log / diff / tree / blame の責務を物理ファイルに分離。public API を完全保持
decision(rpc-dispatcher): handler 群を機能別 extension 11 ファイルに分離。stored properties を private から internal に降格して同 module 内 cross-file extension を許可
decision(gozdcore-layout): Sources/GozdCore/ を Claude / Fs / Git / Pty / Rpc / Shell / Socket / Store / Voicevox のサブディレクトリに分割
decision(gozd-app): GozdApp.swift から AppRuntime / WebPagePush / RpcSchemeHandler / BundleAssetSchemeHandler / PtyHarnessHTML を別ファイルに分離
decision(stderr-direct-check): check-stderr-direct.sh の許可 regex に `(.*/)?` を挟み、許可ファイルが任意のサブディレクトリに置かれていてもマッチするよう書き換え
rejected(rpc-keep-private): stored properties を private のまま保ち handler のみ extension 分離する案は不可。Swift の同型 private は同一ファイル内 extension にのみ許可されるため cross-file 分割は internal 降格が前提
rejected(rpc-handler-out-of-actor): handler を actor 外の関数に外出しする案は却下。actor が保証する RPC serialization を失い、PTYRegistry 等の共有 state mutation が race する
rejected(flat-layout): GozdCore/ を 46 ファイルのフラット 1 ディレクトリで維持する案は却下。Apple 公式 / swift-nio / Vapor 等の大規模 Swift OSS もすべてサブディレクトリで責務分離している
constraint(swift-extension-private): Swift 6.x の private はファイル境界を持つため cross-file extension は internal 以上が必須
learned(swiftpm): Sources/<target>/ 配下のサブディレクトリは Package.swift で path/exclude 指定なしに自動収集される (SwiftPM 5.3+)
learned(stderr-direct-check): 許可リスト regex を絶対パスで書くと、ファイル再配置のたびに更新漏れで pre-commit が破綻する。`(.*/)?` で配置場所に依存しない形にする
