+++ apps/native/Sources/GozdCore/Claude/ClaudeHooksSettings.swift
