+++ apps/native/Sources/GozdCore/Claude/ClaudeSessionLog.swift
