+++ apps/native/Sources/GozdCore/Fs/FSWatchRegistry.swift
