+++ apps/native/Sources/GozdCore/Fs/FSWatcher.swift
