+++ apps/native/Sources/GozdCore/Git/WorktreeOps.swift
