+++ apps/native/Sources/GozdCore/Pty/PTYManager.swift
