+++ apps/native/Sources/GozdCore/Pty/PTYRegistry.swift
