+++ apps/native/Sources/GozdCore/Pty/PTYTrace.swift
