+++ apps/native/Sources/GozdCore/Shell/GozdEnvOverlay.swift
