+++ apps/native/Sources/GozdCore/Shell/ShellCommandOps.swift
