+++ apps/native/Sources/GozdCore/Socket/SocketServer.swift
