+++ apps/native/Sources/GozdCore/Store/AppConfigStore.swift
