+++ apps/native/Sources/GozdCore/Store/AppStateStore.swift
