+++ apps/native/Sources/GozdCore/Store/ProjectKey.swift
